Window bug fix
SS bug fix on first time loading causing breakage of window and handle anim.

Explanation:  Window handle must open = 1 for window to be movable.  Window handle can not move unless window is closed = 0   Although the config states the handle is open = 1 and window is open = 16384, it's not loading correctly.  I'm guessing due the anims not being setup for this.  Defaulting everything to closed = 0  as a temp fix.  Will revisit when anims are redone.   Will also recode later so that the window can still be moved if the handle is down, but won't close all the way because of the handle being in the way :)