Updates
***WARNING*** There are edits in the multiple panel.cfg files of all the planes that may mess up your side. Ignore these if necessary. Specifically the registration lines and the  texture line for the ADF.

Main update: All about the ADF. Between this and the last update, has been heavily modified.  Custom needle anim, this allows ANT mode test, "noise" from night effect and now distance. The whole stock system of active and standby has been thrown out the window due to lack of variables to have control. Instead of ADF1 active and standby, using ADF1 active and ADF2 as the standby. This gives needed control for state saving and for the actvie decimal auto tuning.

*.flt file tweaking. Some localvars removed or changed as other ways have presented themselves.
Some of the <Update> codes have been moved to another file just to organize things better. very WIP for my procrastinating OCD lol

Moved DME and AP to their own circuits using the right side breaks.  I looked at using "visibility code" to hide them but it left a hole in the panel, so I figure we just use as needed and inop the others?

New glass textures to check out??? Not squeaky clean, but just a hint of marks/scratches.

Added new instrument camera view that looks down between the seats to see the ipad and the fuel valve.

OH...this was in the last commit, maybe a new way to go on the copilot vis.  Since there doesn't seem to be a way to have this without having the weight in the flightmodel.cfg...I added and update code to set that station weight back to 0.  I've added back the old viz code for the pilot/copilot (slightly tweaked) So what happens now is we have control over the copilot viz, and its now technically back to 0 and NOT showing the copilot.  The downside is this, and ANY external weight changes really (fuel, pilots, cargo) do not reflect in the sims "Fuel" page. That's not just us, all third party planes trying to get al fancy suffer from the same thing, which is why they, and we, specify to make these changes in the EFB.